Refactor photo review to make it reusuable across mobile view
